refactor(cli): share diagnostic rendering
