+++ cmd/thriftfmt/cli.go
@@ -10,6 +10,7 @@ import (
 	"strconv"
 	"strings"
 
+	cliutil "github.com/kpumuk/thrift-weaver/internal/cli"
 	"github.com/kpumuk/thrift-weaver/internal/format"
 	"github.com/kpumuk/thrift-weaver/internal/syntax"
 	"github.com/kpumuk/thrift-weaver/internal/text"
@@ -214,181 +215,15 @@ func handleRangeResult(stdout, stderr io.Writer, opts cliOptions, original []byt
 
 func handleFormatError(stderr io.Writer, tree *syntax.Tree, diags []syntax.Diagnostic, err error) int {
 	if format.IsErrUnsafeToFormat(err) {
-		writeDiagnostics(stderr, tree, diags)
+		cliutil.WriteDiagnostics(stderr, "thriftfmt", tree, diags, diagnosticDisplayText)
 		writef(stderr, "thriftfmt: %v\n", err)
 		return exitUnsafe
 	}
 	if len(diags) > 0 {
-		writeDiagnostics(stderr, tree, diags)
+		cliutil.WriteDiagnostics(stderr, "thriftfmt", tree, diags, diagnosticDisplayText)
 	}
 	writef(stderr, "thriftfmt: %v\n", err)
 	return exitInternal
-}
-
-func writeDiagnostics(w io.Writer, tree *syntax.Tree, diags []syntax.Diagnostic) {
-	if len(diags) == 0 {
-		return
-	}
-	var li *text.LineIndex
-	uri := ""
-	if tree != nil {
-		li = tree.LineIndex
-		uri = tree.URI
-	}
-	for i, d := range diags {
-		if i > 0 {
-			writeln(w)
-		}
-		prefix := "thriftfmt"
-		if uri != "" {
-			prefix = uri
-		}
-		writeDiagnosticHeader(w, prefix, li, d)
-		writeDiagnosticSnippet(w, tree, li, d)
-	}
-}
-
-func writeDiagnosticHeader(w io.Writer, prefix string, li *text.LineIndex, d syntax.Diagnostic) {
-	loc := d.Span.String()
-	if li != nil && d.Span.Start.IsValid() {
-		if p, err := li.OffsetToPoint(d.Span.Start); err == nil {
-			loc = fmt.Sprintf("%d:%d", p.Line+1, p.Column+1)
-		}
-	}
-	msg, detailLines := diagnosticDisplayText(d)
-	writef(
-		w,
-		"%s:%s: %s: %s/%s: %s\n",
-		prefix,
-		loc,
-		diagnosticSeverityLetter(d.Severity),
-		d.Source,
-		d.Code,
-		msg,
-	)
-	for _, line := range detailLines {
-		writef(w, "  %s\n", line)
-	}
-}
-
-func writeDiagnosticSnippet(w io.Writer, tree *syntax.Tree, li *text.LineIndex, d syntax.Diagnostic) {
-	if tree == nil || li == nil || !d.Span.Start.IsValid() {
-		return
-	}
-
-	startPoint, err := li.OffsetToPoint(d.Span.Start)
-	if err != nil {
-		return
-	}
-	lineStart, lineText, ok := sourceLineAt(tree.Source, d.Span.Start)
-	if !ok {
-		return
-	}
-
-	startCol := min(max(int(d.Span.Start-lineStart), 0), len(lineText))
-
-	caretWidth := diagnosticCaretWidth(li, d, startPoint.Line, len(lineText), lineStart)
-	caretPrefix := caretPrefixForLine(lineText, startCol)
-
-	writeln(w, string(lineText))
-	writeString(w, caretPrefix)
-	writeString(w, strings.Repeat("^", caretWidth))
-	writeln(w)
-}
-
-func diagnosticCaretWidth(li *text.LineIndex, d syntax.Diagnostic, startLine int, lineLen int, lineStart text.ByteOffset) int {
-	if lineLen == 0 {
-		return 1
-	}
-	if !d.Span.End.IsValid() || d.Span.End <= d.Span.Start {
-		return 1
-	}
-
-	end := min(d.Span.End, li.SourceLen())
-
-	endPoint, err := li.OffsetToPoint(end)
-	if err != nil {
-		return 1
-	}
-
-	startCol := min(max(int(d.Span.Start-lineStart), 0), lineLen)
-
-	if endPoint.Line != startLine {
-		if startCol >= lineLen {
-			return 1
-		}
-		return lineLen - startCol
-	}
-
-	endCol := endPoint.Column
-	if endCol < startCol {
-		return 1
-	}
-	if endCol > lineLen {
-		endCol = lineLen
-	}
-	if endCol == startCol {
-		return 1
-	}
-	return endCol - startCol
-}
-
-func sourceLineAt(src []byte, off text.ByteOffset) (text.ByteOffset, []byte, bool) {
-	if !off.IsValid() {
-		return 0, nil, false
-	}
-	i := int(off)
-	if i < 0 || i > len(src) {
-		return 0, nil, false
-	}
-
-	start := i
-	for start > 0 && src[start-1] != '\n' {
-		start--
-	}
-
-	end := i
-	for end < len(src) && src[end] != '\n' {
-		end++
-	}
-	if end > start && src[end-1] == '\r' {
-		end--
-	}
-
-	return text.ByteOffset(start), src[start:end], true
-}
-
-func caretPrefixForLine(line []byte, col int) string {
-	if col <= 0 {
-		return ""
-	}
-	if col > len(line) {
-		col = len(line)
-	}
-
-	var b strings.Builder
-	b.Grow(col)
-	for _, ch := range line[:col] {
-		if ch == '\t' {
-			b.WriteByte('\t')
-			continue
-		}
-		b.WriteByte(' ')
-	}
-	return b.String()
-}
-
-func diagnosticSeverityLetter(s syntax.Severity) string {
-	switch s {
-	case syntax.SeverityError:
-		return "E"
-	case syntax.SeverityWarning:
-		return "W"
-	case syntax.SeverityInfo:
-		return "I"
-	default:
-		return "E"
-	}
 }
 
 func diagnosticDisplayText(d syntax.Diagnostic) (message string, detailLines []string) {
@@ -493,12 +328,10 @@ func writef(w io.Writer, format string, args ...any) {
 }
 
 func writeln(w io.Writer, args ...any) {
-	//nolint:gosec // Terminal/debug output helper.
 	_, _ = fmt.Fprintln(w, args...)
 }
 
 func writeString(w io.Writer, s string) {
-	//nolint:gosec // Terminal/debug output helper.
 	_, _ = io.WriteString(w, s)
 }
